stock control

+++ app/controllers/orders_controller.rb
@@ -22,42 +22,46 @@ class OrdersController < ApplicationController
     product = Product.find(params[:product_id])
     existing_order_item = current_order.order_items.to_a.find { |i| i.product.id === product.id }
 
-    if existing_order_item.present?
-      existing_order_item.price = product.price
-      existing_order_item.weight = product.weight
-      existing_order_item.quantity += 1
-      existing_order_item.save!
-    else
-      current_order_item = current_order.order_items.create!
-      current_order_item.price = product.price
-      current_order_item.weight = product.weight
-      current_order_item.quantity = 1
-      current_order_item.product = product
-      current_order_item.save!
+    if not product.limited or product.inventory > 0
+      if existing_order_item.present?
+        existing_order_item.price = product.price
+        existing_order_item.weight = product.weight
+        existing_order_item.quantity += 1
+        existing_order_item.save!
+      else
+        current_order_item = current_order.order_items.create!
+        current_order_item.price = product.price
+        current_order_item.weight = product.weight
+        current_order_item.quantity = 1
+        current_order_item.product = product
+        current_order_item.save!
+      end
+
+      if current_order.save
+        respond_to do |format|
+          format.html {
+            flash[:info] = I18n.t(:add_product_ok, scope: :edit_order)
+            redirect_to request.referrer
+          }
+
+          format.json {
+            render :json => { :status => :ok }, :status => :ok
+          }
+        end
+      end
+
+      return
     end
 
-    if current_order.save
-      respond_to do |format|
-        format.html {
-          flash[:info] = I18n.t(:add_product_ok, scope: :edit_order)
-          redirect_to request.referrer
-        }
+    respond_to do |format|
+      format.html {
+        flash[:error] = I18n.t(:add_product_ko, scope: :edit_order)
+        redirect_to request.referrer
+      }
 
-        format.json {
-          render :json => { :status => :ok }, :status => :ok
-        }
-      end
-    else
-      respond_to do |format|
-        format.html {
-          flash[:error] = I18n.t(:add_product_ko, scope: :edit_order)
-          redirect_to request.referrer
-        }
-
-        format.json {
-          render :json => { :status => :ko }, :status => :unprocessable_entity
-        }
-      end
+      format.json {
+        render :json => { :status => :ko }, :status => :unprocessable_entity
+      }
     end
   end
 
@@ -65,20 +69,44 @@ class OrdersController < ApplicationController
     products = params.keys.select { |key| /^cart-quantity-of-product-[a-z0-9]+$/.match(key) }
     existing_order_items = current_order.order_items.to_a;
 
+    all_available = true;
+
     products.each do |p|
       quantity = params[p].to_i
       product_id = p.sub(/cart-quantity-of-product-/, '')
-      order_item = existing_order_items.find { |i| i.product.id === product_id }
 
-      if quantity > 0
-        order_item.quantity = quantity
+      product = Product.find(product_id)
+
+      if all_available && quantity >= 0 && ( not product.limited or product.inventory >= quantity )
+        all_available = true
       else
-        current_order.order_items.delete(order_item)
+        all_available = false
+        break
       end
-
-      order_item.save!
     end
 
+    if all_available
+      products.each do |p|
+        quantity = params[p].to_i
+        product_id = p.sub(/cart-quantity-of-product-/, '')
+        order_item = existing_order_items.find { |i| i.product.id === product_id }
+
+        if quantity > 0
+          order_item.quantity = quantity
+        else
+          current_order.order_items.delete(order_item)
+        end
+
+        if order_item.save
+          flash[:success] = I18n.t(:adjust_product_ok, scope: :edit_order)
+          redirect_to manage_cart_orders_path
+
+          return
+        end
+      end
+    end
+
+    flash[:error] = I18n.t(:adjust_product_ko, scope: :edit_order)
     redirect_to manage_cart_orders_path
   end
 
@@ -87,7 +115,20 @@ class OrdersController < ApplicationController
     current_order.user = current_user
     current_order.delivery_destination = current_user.addresses.detect { |a| a.id.to_s == params[:delivery_destination_id] }
 
-    if current_order.save
+    all_available = true;
+
+    current_order.order_items.each do |oi|
+      product = oi.product
+
+      if all_available && ( not product.limited or product.inventory >= oi.quantity )
+        all_available = true
+      else
+        all_available = false
+        break
+      end
+    end
+
+    if all_available && current_order.save
       session.delete(:order_id)
 
       respond_to do |format|
